% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_class_02_kRp.hyphen.R
\docType{class}
\name{kRp.hyphen,-class}
\alias{kRp.hyphen,-class}
\alias{kRp_hyphen}
\alias{kRp.hyphen-class}
\title{S4 Class kRp.hyphen}
\description{
This class is used for objects that are returned by \code{\link[sylly:hyphen]{hyphen}}.
}
\section{Slots}{

\describe{
\item{\code{lang}}{A character string,
      naming the language that is assumed for the analized text in this object}

\item{\code{desc}}{Descriptive statistics of the analyzed text.}

\item{\code{hyphen}}{A data.frame with two columns:
\describe{
  \item{\code{syll}:}{Number of recognized syllables}
  \item{\code{word}:}{The hyphenated word}
}}
}}

\section{Contructor function}{

Should you need to manually generate objects of this class (which should rarely be the case),
      the contructor function 
\code{kRp_hyphen(...)} can be used instead of
\code{new("kRp.hyphen", ...)}. Whenever possible, stick to
\code{\link[sylly:hyphen]{hyphen}}.
}

\keyword{classes}
