% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_method_kRp.hyphen.R
\docType{methods}
\name{describe}
\alias{describe}
\alias{describe,kRp.hyphen-method}
\alias{describe<-}
\alias{describe<-,kRp.hyphen-method}
\alias{language}
\alias{language,kRp.hyphen-method}
\alias{language<-}
\alias{language<-,kRp.hyphen-method}
\alias{hyphenText}
\alias{hyphenText,kRp.hyphen-method}
\alias{hyphenText,-methods}
\alias{hyphenText<-}
\alias{hyphenText<-,kRp.hyphen-method}
\alias{hyphenText<-,-methods}
\alias{[,kRp.hyphen-method}
\alias{[,-methods}
\alias{[,kRp.hyphen,ANY,ANY-method}
\alias{[<-,kRp.hyphen-method}
\alias{[<-,-methods}
\alias{[<-,kRp.hyphen,ANY,ANY,ANY-method}
\alias{[[,kRp.hyphen-method}
\alias{[[,-methods}
\alias{[[,kRp.hyphen,ANY-method}
\alias{[[<-,kRp.hyphen-method}
\alias{[[<-,-methods}
\alias{[[<-,kRp.hyphen,ANY,ANY-method}
\title{Getter/setter methods for sylly objects}
\usage{
describe(obj)

\S4method{describe}{kRp.hyphen}(obj)

describe(obj) <- value

\S4method{describe}{kRp.hyphen}(obj) <- value

language(obj)

\S4method{language}{kRp.hyphen}(obj)

language(obj) <- value

\S4method{language}{kRp.hyphen}(obj) <- value

hyphenText(obj)

\S4method{hyphenText}{kRp.hyphen}(obj)

hyphenText(obj) <- value

\S4method{hyphenText}{kRp.hyphen}(obj) <- value

\S4method{[}{kRp.hyphen}(x, i, j)

\S4method{[}{kRp.hyphen}(x, i, j) <- value

\S4method{[[}{kRp.hyphen}(x, i)

\S4method{[[}{kRp.hyphen}(x, i) <- value
}
\arguments{
\item{obj}{An object of class \code{\link[sylly:kRp.hyphen-class]{kRp.hyphen}}.}

\item{value}{A value to set.}

\item{x}{An object of class \code{\link[sylly:kRp.hyphen-class]{kRp.hyphen}}.}

\item{i}{Row index.}

\item{j}{Column index.}
}
\description{
These methods should be used to get or set values of hyphenated text objects
generated by functions like \code{hyphen()}.
}
\details{
\itemize{
  \item{\code{describe()} }{returns the \code{desc} slot.}
  \item{\code{language()} }{returns the \code{lang} slot.}
  \item{\code{hyphenText()} }{returns the \code{hyphen} slot from objects of class \code{kRp.hyphen}.}
  \item{\code{[}/\code{[[} }{Can be used as a shortcut to index the results of \code{hyphenText()}.}
}
}
\examples{
\dontrun{
hyphenText(hyphenated.txt)
}
}
