% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_method_show.kRp.hyphen.R
\docType{methods}
\name{show,kRp.hyphen-method}
\alias{show,kRp.hyphen-method}
\alias{show}
\title{Show method for sylly objects}
\usage{
\S4method{show}{kRp.hyphen}(object)
}
\arguments{
\item{object}{An object of class \code{kRp.hyphen}.}
}
\description{
Show method for S4 objects of class \code{\link[sylly:kRp.hyphen-class]{kRp.hyphen}}.
}
\examples{
\dontrun{
  hyphen(tagged.text)
}
}
\seealso{
\code{\link[sylly:kRp.hyphen-class]{kRp.hyphen}}
}
\keyword{methods}
