% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_method_summary.kRp.hyphen.R
\docType{methods}
\name{summary,kRp.hyphen-method}
\alias{summary,kRp.hyphen-method}
\alias{summary}
\title{Summary method for sylly objects}
\usage{
\S4method{summary}{kRp.hyphen}(object)
}
\arguments{
\item{object}{An object of class \code{kRp.hyphen}.}
}
\description{
Summary method for S4 objects of class \code{\link[sylly:kRp.hyphen-class]{kRp.hyphen}}.
}
\examples{
\dontrun{
summary(hyphen(tagged.text))
}
}
\seealso{
\code{\link[sylly:kRp.hyphen-class]{kRp.hyphen}}
}
\keyword{methods}
