% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sylly-package.R
\docType{package}
\name{sylly-package}
\alias{sylly}
\alias{sylly-package}
\title{\packageTitle{sylly}}
\description{
\packageDescription{sylly}
}
\details{
The DESCRIPTION file:
\tabular{ll}{
Package: \tab sylly\cr
Type: \tab Package\cr
Version: \tab 0.1-5\cr
Date: \tab 2018-07-29\cr
Depends: \tab R (>= 3.0.0)\cr
Encoding: \tab UTF-8\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
URL: \tab https://reaktanz.de/?c=hacking&s=sylly\cr
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://reaktanz.de/?c=hacking&s=sylly}
  \item Report bugs at \url{https://github.com/unDocUMeantIt/sylly/issues}
}

}
\author{
\packageAuthor{sylly}

Maintainer: \packageMaintainer{sylly}
}
